+++ rio_tiler/mosaic/__init__.py
@@ -1,6 +1,14 @@
 """rio-tiler.mosaic."""
 
 from . import methods
+from .backend import AsyncBaseBackend, BaseBackend
 from .reader import async_mosaic_reader, mosaic_point_reader, mosaic_reader
 
-__all__ = ["methods", "async_mosaic_reader", "mosaic_point_reader", "mosaic_reader"]
+__all__ = [
+    "methods",
+    "async_mosaic_reader",
+    "mosaic_point_reader",
+    "mosaic_reader",
+    "BaseBackend",
+    "AsyncBaseBackend",
+]